Add a function to preferentially connect to polygon faces
This is to replicate old THERMOS UI behaviour found in
thermos-importer, where the connector-forming code prefers to connect
to the middle of a wall even if it makes the connector a little bit
longer.

The logic is not quite identical, as thermos-importer considers
multiple nearby paths, whereas we min-by distance before adding the
snap points.