FIX: empty tables would truncate geopackage reads across tables
This was caused by the situation where we can open an iterator for a
table, but the iterator is empty, and the outer cross-table iterator
just passed through to the inner iterator for its (.hasNext) method.

So, even though there were more tables left to read from the outer
iterator looked empty. The fix is that in maybe-advance! where we move
to the next table, we check if we've moved onto an empty table and
skip over it.